@@ -7,7 +7,11 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "compose": "python3 scripts/compose.py"
+    "compose": "python3 scripts/compose.py",
+    "lint": "ng lint --fix",
+    "format": "prettier-eslint --eslint-config-path ./.eslintrc.json --write '**/*.ts'",
+    "e2e": "ng e2e",
+    "prettier": "npx prettier --write ."
   },
   "private": true,
   "dependencies": {
@@ -29,21 +33,33 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.0.5",
+    "@angular-eslint/builder": "14.0.2",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/schematics": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
     "@angular/cli": "^14.0.5",
     "@angular/compiler-cli": "^14.0.5",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
     "angular-cli-ghpages": "^1.0.0",
-    "eslint": "^8.8.0",
+    "eslint": "^8.18.0",
     "eslint-config-google": "^0.14.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-json": "^3.1.0",
+    "eslint-plugin-jsonc": "^2.3.1",
+    "eslint-plugin-prettier": "^4.2.1",
     "jasmine-core": "~3.10.0",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
+    "prettier": "^2.7.1",
+    "prettier-eslint": "^15.0.1",
+    "prettier-eslint-cli": "^6.0.1",
     "typescript": "~4.7.4"
   }
 }
