@@ -1,4 +1,3 @@
-import { Post } from "src/app/modules/blog/models/post"; 
+import { Post } from 'src/app/modules/blog/models/post';
 
-export const reviews: Post[] = [
-]
+export const reviews: Post[] = [];
