@@ -1,24 +1,16 @@
-import { TestBed } from '@angular/core/testing';
-import { Router } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
 import { MenuModule } from './modules/menu/menu.module';
 import { MenuService } from './modules/menu/services/menu.service';
+import { TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 
 const compileComponent = () => {
   TestBed.configureTestingModule({
-    imports: [
-      RouterTestingModule,
-      MenuModule,
-    ],
-    declarations: [
-      AppComponent,
-    ],
-    providers: [
-      MenuService,
-    ]
+    imports: [RouterTestingModule, MenuModule],
+    declarations: [AppComponent],
+    providers: [MenuService],
   }).compileComponents();
-}
+};
 
 describe('AppComponent', () => {
   beforeEach(async () => {
