@@ -1,22 +1,118 @@
-import { Post } from "../modules/blog/models/post";
+import { Post } from '../modules/blog/models/post';
 
 export const GUIDES: Post[] = [
-    new Post('guides', 'Git', 'Git History', 'Lets assume you are building your own application on the codebase of a repository and are interestin...', './src/assets/posts/guides/008_git_history/008_git_history.md', 'assets/posts/guides/008_git_history/thumbnail.png', '2022-06-20'), 
-    new Post('guides', 'Docker', 'Multi Container Applications', 'As we have seen in the previous blog post, it is rather simple to run a Flask application inside a d...', './src/assets/posts/guides/006_multi_container_applications/006_multi_container_applications.md', 'assets/posts/guides/006_multi_container_applications/thumbnail.png', '2022-06-27'), 
-    new Post('guides', 'RxJS', 'DebounceTime', 'The debounce ope....', './src/assets/posts/guides/009_searching_with_rxjs/009_searching_with_rxjs.md', 'assets/posts/guides/009_searching_with_rxjs/thumbnail.png', '2022-07-04'), 
-    new Post('guides', 'Docker', 'Containerize Flask Apps', 'As we have seen in the previous post that Docker provides a rather sophiticated toolset for creating...', './src/assets/posts/guides/005_containerize_flask_applications/005_containerize_flask_applications.md', 'assets/posts/guides/005_containerize_flask_applications/thumbnail.png', '2022-06-27'), 
-    new Post('guides', 'Python', 'Virtual Environments', 'It is considered best practice to run Python applications inside isolated environments different fro...', './src/assets/posts/guides/007_python_virtual_environments/007_python_virtual_environments.md', 'assets/posts/guides/007_python_virtual_environments/thumbnail.png', '2022-06-27'), 
-    new Post('guides', 'RxJS', 'RxJS in Angular', 'RxJS (Reactive Extensions for JavaScript) is a library that enables reactive programming in Angular,...', './src/assets/posts/guides/002_introduction_to_rxjs/introduction_to_rxjs.md', 'assets/posts/guides/002_introduction_to_rxjs/thumbnail.png', '2022-07-04'), 
-    new Post('guides', 'Angular', 'Angular on GitHub Pages', 'GitHub Pages is a free hosting service for static websides offered by Github....', './src/assets/posts/guides/001_angular_apps_on_github_pages/angular_apps_on_github_pages.md', 'assets/posts/guides/001_angular_apps_on_github_pages/thumbnail.png', '2022-07-08'), 
-    new Post('guides', 'RxJS', 'RxJS Operators', 'The core strength of RxJS lies in applying operators on Observables to filter and transform the retu...', './src/assets/posts/guides/003_rxjs_pipe_operators/003_rxjs_pipe_operators.md', 'assets/posts/guides/003_rxjs_pipe_operators/thumbnail.png', '2022-07-04'), 
+  new Post(
+    'guides',
+    'Git',
+    'Git History',
+    'Lets assume you are building your own application on the codebase of a repository and are interestin...',
+    './src/assets/posts/guides/008_git_history/008_git_history.md',
+    'assets/posts/guides/008_git_history/thumbnail.png',
+    '2022-06-20'
+  ),
+  new Post(
+    'guides',
+    'Docker',
+    'Multi Container Applications',
+    'As we have seen in the previous blog post, it is rather simple to run a Flask application inside a d...',
+    './src/assets/posts/guides/006_multi_container_applications/006_multi_container_applications.md',
+    'assets/posts/guides/006_multi_container_applications/thumbnail.png',
+    '2022-06-27'
+  ),
+  new Post(
+    'guides',
+    'RxJS',
+    'DebounceTime',
+    'The debounce ope....',
+    './src/assets/posts/guides/009_searching_with_rxjs/009_searching_with_rxjs.md',
+    'assets/posts/guides/009_searching_with_rxjs/thumbnail.png',
+    '2022-07-04'
+  ),
+  new Post(
+    'guides',
+    'Docker',
+    'Containerize Flask Apps',
+    'As we have seen in the previous post that Docker provides a rather sophiticated toolset for creating...',
+    './src/assets/posts/guides/005_containerize_flask_applications/005_containerize_flask_applications.md',
+    'assets/posts/guides/005_containerize_flask_applications/thumbnail.png',
+    '2022-06-27'
+  ),
+  new Post(
+    'guides',
+    'Python',
+    'Virtual Environments',
+    'It is considered best practice to run Python applications inside isolated environments different fro...',
+    './src/assets/posts/guides/007_python_virtual_environments/007_python_virtual_environments.md',
+    'assets/posts/guides/007_python_virtual_environments/thumbnail.png',
+    '2022-06-27'
+  ),
+  new Post(
+    'guides',
+    'RxJS',
+    'RxJS in Angular',
+    'RxJS (Reactive Extensions for JavaScript) is a library that enables reactive programming in Angular,...',
+    './src/assets/posts/guides/002_introduction_to_rxjs/introduction_to_rxjs.md',
+    'assets/posts/guides/002_introduction_to_rxjs/thumbnail.png',
+    '2022-07-04'
+  ),
+  new Post(
+    'guides',
+    'Angular',
+    'Angular on GitHub Pages',
+    'GitHub Pages is a free hosting service for static websides offered by Github....',
+    './src/assets/posts/guides/001_angular_apps_on_github_pages/angular_apps_on_github_pages.md',
+    'assets/posts/guides/001_angular_apps_on_github_pages/thumbnail.png',
+    '2022-07-08'
+  ),
+  new Post(
+    'guides',
+    'RxJS',
+    'RxJS Operators',
+    'The core strength of RxJS lies in applying operators on Observables to filter and transform the retu...',
+    './src/assets/posts/guides/003_rxjs_pipe_operators/003_rxjs_pipe_operators.md',
+    'assets/posts/guides/003_rxjs_pipe_operators/thumbnail.png',
+    '2022-07-04'
+  ),
 ];
 
 export const RXJS_RELATED_POSTS: Post[] = [
-    new Post('guides', 'RxJS', 'DebounceTime', 'The debounce ope....', './src/assets/posts/guides/009_searching_with_rxjs/009_searching_with_rxjs.md', 'assets/posts/guides/009_searching_with_rxjs/thumbnail.png', '2022-07-04'), 
-    new Post('guides', 'RxJS', 'RxJS in Angular', 'RxJS (Reactive Extensions for JavaScript) is a library that enables reactive programming in Angular,...', './src/assets/posts/guides/002_introduction_to_rxjs/introduction_to_rxjs.md', 'assets/posts/guides/002_introduction_to_rxjs/thumbnail.png', '2022-07-04'), 
-    new Post('guides', 'RxJS', 'RxJS Operators', 'The core strength of RxJS lies in applying operators on Observables to filter and transform the retu...', './src/assets/posts/guides/003_rxjs_pipe_operators/003_rxjs_pipe_operators.md', 'assets/posts/guides/003_rxjs_pipe_operators/thumbnail.png', '2022-07-04'), 
+  new Post(
+    'guides',
+    'RxJS',
+    'DebounceTime',
+    'The debounce ope....',
+    './src/assets/posts/guides/009_searching_with_rxjs/009_searching_with_rxjs.md',
+    'assets/posts/guides/009_searching_with_rxjs/thumbnail.png',
+    '2022-07-04'
+  ),
+  new Post(
+    'guides',
+    'RxJS',
+    'RxJS in Angular',
+    'RxJS (Reactive Extensions for JavaScript) is a library that enables reactive programming in Angular,...',
+    './src/assets/posts/guides/002_introduction_to_rxjs/introduction_to_rxjs.md',
+    'assets/posts/guides/002_introduction_to_rxjs/thumbnail.png',
+    '2022-07-04'
+  ),
+  new Post(
+    'guides',
+    'RxJS',
+    'RxJS Operators',
+    'The core strength of RxJS lies in applying operators on Observables to filter and transform the retu...',
+    './src/assets/posts/guides/003_rxjs_pipe_operators/003_rxjs_pipe_operators.md',
+    'assets/posts/guides/003_rxjs_pipe_operators/thumbnail.png',
+    '2022-07-04'
+  ),
 ];
 
 export const PRODUCTIVITY: Post[] = [
-    new Post('productivity', 'Routines', 'On Routines', 'Regardless of what we pursue, having a set of highly effective routines in place is crucial for comi...', './src/assets/posts/productivity/202_daily_routines/202_daily_routines.md', 'assets/posts/productivity/202_daily_routines/thumbnail.png', '2022-07-21'), 
+  new Post(
+    'productivity',
+    'Routines',
+    'On Routines',
+    'Regardless of what we pursue, having a set of highly effective routines in place is crucial for comi...',
+    './src/assets/posts/productivity/202_daily_routines/202_daily_routines.md',
+    'assets/posts/productivity/202_daily_routines/thumbnail.png',
+    '2022-07-21'
+  ),
 ];
