@@ -1,6 +1,6 @@
 import { Component, HostListener, OnInit } from '@angular/core';
 import { ActivatedRoute, ParamMap, Router } from '@angular/router';
-import { map, Observable, of, tap } from 'rxjs';
+import { Observable, map, of, tap } from 'rxjs';
 import { Post } from '../../models/post';
 import { BlogService } from '../../services/blog.service';
 import { WindowService } from '../../services/window.service';
@@ -8,31 +8,36 @@ import { WindowService } from '../../services/window.service';
 @Component({
   selector: 'app-blog',
   templateUrl: './blog.component.html',
-  styleUrls: ['./blog.component.scss']
+  styleUrls: ['./blog.component.scss'],
 })
 export class BlogComponent implements OnInit {
   numberOfColumns: number = this.windowService.getNumberOfColumns();
-  posts: Observable<Post[]> = this.fetchPosts('guides')
+  posts: Observable<Post[]> = this.fetchPosts('guides');
 
   constructor(
     private readonly blogService: BlogService,
     private readonly router: Router,
     private readonly activatedRoute: ActivatedRoute,
-    private readonly windowService: WindowService,
+    private readonly windowService: WindowService
   ) {}
 
   ngOnInit(): void {
-    this.activatedRoute.paramMap.pipe(
-      tap(console.log),
-      map((paramMap: ParamMap) => [paramMap.get('category'), paramMap.get('topic')]),
-      tap(console.log),
-      tap((params) => this.posts = this.fetchPosts(params[0], params[1])),
-    ).subscribe();
+    this.activatedRoute.paramMap
+      .pipe(
+        tap(console.log),
+        map((paramMap: ParamMap) => [
+          paramMap.get('category'),
+          paramMap.get('topic'),
+        ]),
+        tap(console.log),
+        tap(params => (this.posts = this.fetchPosts(params[0], params[1])))
+      )
+      .subscribe();
   }
 
   @HostListener('window:resize', ['$event'])
-  public onResize(event: { target: { innerWidth: number; }; }): void {
-    this.numberOfColumns = this.onWidthChange(event.target.innerWidth)
+  public onResize(event: { target: { innerWidth: number } }): void {
+    this.numberOfColumns = this.onWidthChange(event.target.innerWidth);
   }
 
   private onWidthChange(width: number): number {
@@ -43,19 +48,24 @@ export class BlogComponent implements OnInit {
   }
 
   public showPost(link: string): void {
-    const filePath = link.replace('./src/assets/posts', '')
+    const filePath = link.replace('./src/assets/posts', '');
     this.router.navigate([`blog/${filePath}`]);
   }
 
   public selectTopic(topic: string): void {
-    this.activatedRoute.paramMap.pipe(
-      map((paramMap: ParamMap) => paramMap.get('category')),
-      map((category) => !category ? 'guides' : category),
-      tap((category) => this.router.navigate([`blog/${category}/${topic}`]))
-    ).subscribe();
+    this.activatedRoute.paramMap
+      .pipe(
+        map((paramMap: ParamMap) => paramMap.get('category')),
+        map(category => (!category ? 'guides' : category)),
+        tap(category => this.router.navigate([`blog/${category}/${topic}`]))
+      )
+      .subscribe();
   }
 
-  private fetchPosts(category: string | null, topic?: string | null): Observable<Post[]> {
+  private fetchPosts(
+    category: string | null,
+    topic?: string | null
+  ): Observable<Post[]> {
     category = category !== null ? category : 'guides';
     topic = topic !== null ? topic : 'All';
     return of(this.blogService.getPosts(category, topic));
@@ -64,7 +74,10 @@ export class BlogComponent implements OnInit {
   selectCategory(): Observable<string> {
     return this.activatedRoute.paramMap.pipe(
       map((paramMap: ParamMap) => paramMap.get('category') ?? 'Guides'),
-      map((category: string) => category.charAt(0).toUpperCase() + category.slice(1)),
+      map(
+        (category: string) =>
+          category.charAt(0).toUpperCase() + category.slice(1)
+      )
     );
   }
 }
