@@ -1,22 +1,19 @@
+import { PostComponent } from './post.component';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { PostComponent } from './post.component';
 
 const compileComponent = () => {
   TestBed.configureTestingModule({
-    imports: [
-      RouterTestingModule,
-    ],
-    providers: [
-    ],
-    teardown: { destroyAfterEach: false }
+    imports: [RouterTestingModule],
+    providers: [],
+    teardown: { destroyAfterEach: false },
   }).compileComponents();
-}
+};
 
 describe('PostComponent', () => {
   let component: PostComponent;
   let fixture: ComponentFixture<PostComponent>;
-  
+
   beforeEach(async () => {
     compileComponent();
   });
