@@ -1,22 +1,25 @@
-import { TestBed } from '@angular/core/testing';
-import { GUIDES, PRODUCTIVITY, RXJS_RELATED_POSTS } from 'src/app/helpers/post-mocks';
 import { Post } from '../models/post';
 import { PostCollection } from '../models/postCollection';
 import { BlogService } from './blog.service';
-
+import { TestBed } from '@angular/core/testing';
+import {
+  GUIDES,
+  PRODUCTIVITY,
+  RXJS_RELATED_POSTS,
+} from 'src/app/helpers/post-mocks';
 
 const postCollection = jasmine.createSpyObj('PostCollection', ['selectPosts']);
-postCollection.selectPosts.withArgs('guides').and.returnValue(GUIDES)
-postCollection.selectPosts.withArgs('productivity').and.returnValue(PRODUCTIVITY)
+postCollection.selectPosts.withArgs('guides').and.returnValue(GUIDES);
+postCollection.selectPosts
+  .withArgs('productivity')
+  .and.returnValue(PRODUCTIVITY);
 
 const compileComponent = () => {
   TestBed.configureTestingModule({
-      providers: [
-        { provide: PostCollection, useValue: postCollection }
-      ],
-      teardown: { destroyAfterEach: false }
+    providers: [{ provide: PostCollection, useValue: postCollection }],
+    teardown: { destroyAfterEach: false },
   }).compileComponents();
-}
+};
 
 describe('BlogService', () => {
   let service: BlogService;
@@ -30,14 +33,14 @@ describe('BlogService', () => {
     expect(service).toBeTruthy();
   });
 
-  fit('should retrieve all posts belonging to the productivity category', () => {
-    const result: Post[] = service.getPosts('productivity')
-    expect(result).toEqual(PRODUCTIVITY)
+  it('should retrieve all posts belonging to the productivity category', () => {
+    const result: Post[] = service.getPosts('productivity');
+    expect(result).toEqual(PRODUCTIVITY);
   });
 
   it('should only retrieve posts about RxJS', () => {
-    const result: Post[] = service.getPosts('guides', 'RxJS')
-    expect(result).toEqual(RXJS_RELATED_POSTS)
+    const result: Post[] = service.getPosts('guides', 'RxJS');
+    expect(result).toEqual(RXJS_RELATED_POSTS);
   });
 
   it('should return all topics belonging to "guides" category', () => {
