@@ -1,19 +1,16 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { APP_BASE_HREF } from '@angular/common';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
-import { Router } from '@angular/router';
-import { MenuComponent } from './menu.component';
-import { MenuService } from '../../services/menu.service';
-import { RouterTestingModule } from '@angular/router/testing';
-import { MatMenuModule } from '@angular/material/menu';
 import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
+import { RouterTestingModule } from '@angular/router/testing';
+import { MenuService } from '../../services/menu.service';
+import { MenuComponent } from './menu.component';
 
 const compileComponent = () => {
   TestBed.configureTestingModule({
-    declarations: [
-      MenuComponent,
-    ],
+    declarations: [MenuComponent],
     imports: [
       FormsModule,
       RouterTestingModule,
@@ -21,13 +18,9 @@ const compileComponent = () => {
       MatIconModule,
       MatToolbarModule,
     ],
-    providers: [
-      MenuService,
-      { provide: APP_BASE_HREF, useValue : '/' }
-    ],
-  })
-  .compileComponents();
-}
+    providers: [MenuService, { provide: APP_BASE_HREF, useValue: '/' }],
+  }).compileComponents();
+};
 
 describe('MenuComponent', () => {
   let component: MenuComponent;
