@@ -1,7 +1,7 @@
 export class MenuItem {
-    public constructor(
-        public name: string,
-        public icon: string,
-        public link: string,
-    ) {}
+  public constructor(
+    public name: string,
+    public icon: string,
+    public link: string
+  ) {}
 }
