@@ -1,5 +1,5 @@
 import { Post } from "src/app/modules/blog/models/post"; 
 
 export const productivity: Post[] = [
-    new Post('productivity', 'Routines', 'On Routines', 'Regardless of what we pursue, having a set of highly effective routines in place is crucial for comi...', './src/assets/posts/productivity/202_daily_routines/202_daily_routines.md', 'assets/posts/productivity/202_daily_routines/thumbnail.png', '2022-07-21'), 
+    new Post('productivity', 'Routines', 'Healthy Routines', 'Having a set of healthy routines in place is crucial for living a more balanced life while achieving higher golas....', './src/assets/posts/productivity/202_daily_routines/202_daily_routines.md', 'assets/posts/productivity/202_daily_routines/thumbnail.png', '2022-07-21'), 
 ]
