@@ -45,8 +45,9 @@
 /***************************************************************************************************
  * Zone JS is required by default for Angular itself.
  */
-import 'zone.js';  // Included with Angular CLI.
+import 'zone.js';
 
+// Included with Angular CLI.
 
 /***************************************************************************************************
  * APPLICATION IMPORTS
